Introducing MD5 file integrity checks and fixes

+++ src/api/gateway/main.py
@@ -5,8 +5,8 @@ from gateway.database import get_db
 from gateway.auth import verify_jwt, verify_api_key, verify_admin_jwt, verify_admin_api_key
 from gateway.models import Recording, Fragment
 from gateway.schemas import RecordingCreate, FragmentCreate
-from gateway.file_manager import save_fragment_file, get_fragment_file, delete_fragment_file
-from gateway.utils import get_fragment_in_recording, generate_api_key, generate_api_secret
+from gateway.file_manager import delete_recording_files, delete_user_directory, save_fragment_file, get_fragment_file_path, delete_fragment_file, verify_md5, file_response_with_md5
+from gateway.utils import get_admin_recording, get_fragment_in_recording, generate_api_key, generate_api_secret
 
 import os
 import uvicorn
@@ -32,6 +32,7 @@ async def create_fragment(
     index: str = Form(...),
     sample_number: str = Form(...),
     file: UploadFile = File(...),
+    md5_checksum: str = Form(None),
     db: Session = Depends(get_db),
     user = Depends(verify_jwt),
     api_key: str = Depends(verify_api_key)
@@ -40,6 +41,7 @@ async def create_fragment(
     recording = db.query(Recording).filter_by(id=recording_id, owner_id=user_id).first()
     if not recording:
         raise HTTPException(status_code=404, detail="Recording not found")
+    verify_md5(file.file, md5_checksum)
     save_fragment_file(user_id, recording_id, index, recording.file_extension, file)
     new_fragment = Fragment(recording_id=recording_id, index=int(index), sample_number=int(sample_number))
     db.add(new_fragment)
@@ -84,7 +86,7 @@ async def get_fragment_metadata(
 
     return fragment
 
-@app.get("/recordings/{recording_id}/fragments/{index}/file")
+""" @app.get("/recordings/{recording_id}/fragments/{index}/file")
 async def get_fragment_file(
     recording_id: str,
     index: int,
@@ -95,7 +97,7 @@ async def get_fragment_file(
     user_id = user["sub"]
     _, recording = get_fragment_in_recording(db, recording_id, index, user_id)
     
-    file_path = get_fragment_file(
+    file_path = get_fragment_file_path(
         user_id=user_id,
         recording_id=recording_id,
         index=index,
@@ -105,7 +107,7 @@ async def get_fragment_file(
     return FileResponse(
         file_path,
         filename=f"{index}.{recording.file_extension}",
-    )
+    ) """
 
 @app.delete("/recordings/{recording_id}")
 async def delete_recording(recording_id: str, db: Session = Depends(get_db), user = Depends(verify_jwt), api_key: str = Depends(verify_api_key)):
@@ -118,8 +120,7 @@ async def delete_recording(recording_id: str, db: Session = Depends(get_db), use
     )
     if not recording:
         raise HTTPException(status_code=404, detail="Recording not found")
-    for fragment in recording.fragments:
-        delete_fragment_file(user_id, recording_id, fragment.index, recording.file_extension)
+    delete_recording_files(user_id, recording_id)
     db.delete(recording)
     db.commit()
     return {"message": "Recording and associated fragments deleted"}
@@ -171,18 +172,55 @@ async def admin_get_recording(recording_id: str, db: Session = Depends(get_db), 
         raise HTTPException(status_code=404, detail="Recording not found")
     return recording
 
-@app.get("/{admin_url}/recordings/{recording_id}/fragments/{index}")
-async def admin_get_fragment_metadata(recording_id: str, index: int, db: Session = Depends(get_db), user: dict = Depends(verify_admin_jwt), api_key: str = Depends(verify_admin_api_key)):
-    """ Returns metadata of a specific fragment in specific recording from the database """
-    fragment, _ = get_fragment_in_recording(db, recording_id, index, user["sub"])
-    return fragment
-
 @app.get("/{admin_url}/recordings/{recording_id}/fragments/{index}/file")
-async def admin_get_fragment_file(recording_id: str, index: int, db: Session = Depends(get_db), user: dict = Depends(verify_admin_jwt), api_key: str = Depends(verify_admin_api_key)):
-    """ Returns file of a specific fragment in specific recording from the database """
-    _, recording = get_fragment_in_recording(db, recording_id, index, user["sub"])
-    file_path = get_fragment_file(user["sub"], recording_id, index, recording.file_extension)
-    return FileResponse(file_path, filename=f"{index}.{recording.file_extension}")
+async def admin_get_fragment_file(
+    recording_id: str,
+    index: int,
+    db: Session = Depends(get_db),
+    user: dict = Depends(verify_admin_jwt),
+    api_key: str = Depends(verify_admin_api_key)
+):
+    recording = get_admin_recording(db, user, recording_id)
+
+    fragment = (
+        db.query(Fragment)
+        .filter_by(recording_id=recording_id, index=index)
+        .first()
+    )
+    if not fragment:
+        raise HTTPException(status_code=404, detail="Fragment not found")
+
+    file_path = get_fragment_file_path(
+        user_id=recording.owner_id,
+        recording_id=recording_id,
+        index=index,
+        extension=recording.file_extension,
+    )
+    return file_response_with_md5(
+        file_path,
+        filename=f"{index}.{recording.file_extension}",
+    )
+
+@app.get("/recordings/{recording_id}/fragments/{index}/file")
+async def get_fragment_file(
+    recording_id: str,
+    index: int,
+    db: Session = Depends(get_db),
+    user = Depends(verify_jwt),
+    api_key: str = Depends(verify_api_key),
+):
+    user_id = user["sub"]
+    _, recording = get_fragment_in_recording(db, recording_id, index, user_id)
+    file_path = get_fragment_file_path(
+        user_id=user_id,
+        recording_id=recording_id,
+        index=index,
+        extension=recording.file_extension,
+    )
+    return file_response_with_md5(
+        file_path,
+        filename=f"{index}.{recording.file_extension}",
+    )
 
 @app.post("/{admin_url}/users/{owner_id}/recordings", status_code=status.HTTP_201_CREATED)
 async def admin_create_recording(owner_id: str, recording: RecordingCreate, db: Session = Depends(get_db), user: dict = Depends(verify_admin_jwt), api_key: str = Depends(verify_admin_api_key)):
@@ -194,21 +232,31 @@ async def admin_create_recording(owner_id: str, recording: RecordingCreate, db: 
     return new_recording
 
 @app.post("/{admin_url}/users/{owner_id}/recordings/{recording_id}/fragments", status_code=status.HTTP_201_CREATED)
-async def admin_create_fragment(owner_id: str, recording_id: str, fragment: FragmentCreate, db: Session = Depends(get_db), user: dict = Depends(verify_admin_jwt), api_key: str = Depends(verify_admin_api_key)):
-    """ Creates a new fragment in a specific recording for a specific user """
-    recording = db.query(Recording).filter_by(id=fragment.recording_id).first()
+async def admin_create_fragment(
+    owner_id: str,
+    recording_id: str,
+    index: str = Form(...),
+    sample_number: str = Form(...),
+    file: UploadFile = File(...),
+    md5_checksum: str = Form(None),
+    db: Session = Depends(get_db),
+    user: dict = Depends(verify_admin_jwt),
+    api_key: str = Depends(verify_admin_api_key)
+):
+    recording = db.query(Recording).filter_by(id=recording_id, owner_id=owner_id).first()
     if not recording:
         raise HTTPException(status_code=404, detail="Recording not found")
-    save_fragment_file(recording.owner_id, fragment.recording_id, fragment.index, recording.file_extension, fragment.file)
-    new_fragment = Fragment(recording_id=fragment.recording_id, index=int(fragment.index), sample_number=int(fragment.sample_number))
+    verify_md5(file.file, md5_checksum)
+    save_fragment_file(owner_id, recording_id, index, recording.file_extension, file)
+    new_fragment = Fragment(recording_id=recording_id, index=int(index), sample_number=int(sample_number))
     db.add(new_fragment)
     db.commit()
     db.refresh(new_fragment)
     return new_fragment
 
-@app.put("/{admin_url}/users/{owner_id}/recordings/{recording_id}")
-async def admin_update_recording(owner_id: str, recording_id: str, recording: RecordingCreate, db: Session = Depends(get_db), user: dict = Depends(verify_admin_jwt), api_key: str = Depends(verify_admin_api_key)):
-    """ Updates metadata of a specific recording for a specific user """
+@app.put("/{admin_url}/recordings/{recording_id}")
+async def admin_update_recording(recording_id: str, recording: RecordingCreate, db: Session = Depends(get_db), user: dict = Depends(verify_admin_jwt), api_key: str = Depends(verify_admin_api_key)):
+    """ Updates metadata of a specific recording """
     existing_recording = db.query(Recording).filter_by(id=recording_id).first()
     if not existing_recording:
         raise HTTPException(status_code=404, detail="Recording not found")
@@ -218,15 +266,21 @@ async def admin_update_recording(owner_id: str, recording_id: str, recording: Re
     db.refresh(existing_recording)
     return existing_recording
 
-@app.put("/{admin_url}/recordings/{recording_id}/fragments/{index}")
+""" @app.put("/{admin_url}/recordings/{recording_id}/fragments/{index}")
 async def admin_update_fragment(recording_id: str, index: int, fragment: FragmentCreate, db: Session = Depends(get_db), user: dict = Depends(verify_admin_jwt), api_key: str = Depends(verify_admin_api_key)):
-    """ Updates metadata of a specific fragment in a specific recording for a specific user """
-    fragment, _ = get_fragment_in_recording(db, recording_id, index, user["sub"])
+    #Updates metadata of a specific fragment in a specific recording
+    fragment = (
+        db.query(Fragment)
+        .filter_by(recording_id=recording_id, index=index)
+        .first()
+    )
+    if not fragment:
+        raise HTTPException(status_code=404, detail="Fragment not found")
     for key, value in fragment.dict().items():
         setattr(fragment, key, value)
     db.commit()
     db.refresh(fragment)
-    return fragment
+    return fragment """
 
 @app.delete("/{admin_url}/recordings/{recording_id}")
 async def admin_delete_recording(recording_id: str, db: Session = Depends(get_db), user: dict = Depends(verify_admin_jwt), api_key: str = Depends(verify_admin_api_key)):
@@ -234,8 +288,7 @@ async def admin_delete_recording(recording_id: str, db: Session = Depends(get_db
     recording = db.query(Recording).filter_by(id=recording_id).first()
     if not recording:
         raise HTTPException(status_code=404, detail="Recording not found")
-    for fragment in recording.fragments:
-        delete_fragment_file(recording.owner_id, recording_id, fragment.index, recording.file_extension)
+    delete_recording_files(recording.owner_id, recording_id)
     db.delete(recording)
     db.commit()
     return {"message": "Recording and associated fragments deleted"}
@@ -243,7 +296,14 @@ async def admin_delete_recording(recording_id: str, db: Session = Depends(get_db
 @app.delete("/{admin_url}/recordings/{recording_id}/fragments/{index}")
 async def admin_delete_fragment(recording_id: str, index: int, db: Session = Depends(get_db), user: dict = Depends(verify_admin_jwt), api_key: str = Depends(verify_admin_api_key)):
     """ Deletes a specific fragment in a specific recording by recording_id and fragment index """
-    fragment, recording = get_fragment_in_recording(db, recording_id, index, user["sub"])
+    recording = get_admin_recording(db, user, recording_id)
+    fragment = (
+        db.query(Fragment)
+        .filter_by(recording_id=recording_id, index=index)
+        .first()
+    )
+    if not fragment:
+        raise HTTPException(status_code=404, detail="Fragment not found")
     delete_fragment_file(recording.owner_id, recording_id, index, recording.file_extension)
     db.delete(fragment)
     db.commit()
